UnwrapLock + move away from lazy_static
Lazy static uses a spin lock. However most of our statics are also
protected by a lock, meaning that we have a lock around a lock.

UnwrapLock is a unsafe struct that provides a lock around a `MaybeUninit`.
Lock will return the `<T>` within the `MaybeUninit` while a
`lock_uninit` can be used to init the lock.

This is useful for statics that can't be const init but will be
initialized during boot.