fixed static logger static size bug.
I previously added consts for max count of level filters and target
remmappings but I forgot to actually use them.
Because of that size_of<StaticLogger> was so large that the constructor
tried to create a massive stack, which caused some kind of fault.

Not sure why this only happend when we used more than 1 rename