QemuRunner: Rework to use Result and tokio
In order to support isolated tests, panic recovery, etc we need a way to
communicate with the kernel. The kernel listens on serial 2, which we
will connect to using tcp.

This commit does some basic work, to facilitate that future change.
Mainly all panics are reworked to Results and std::process is replaced
with tokio::process.
That way we have an easier way to implement the communication using tcp
in the future.