Add noto-sans-mono font dependency
Unlike other fonts in the future this dependency directly includes the
binary bitmap data for the font in the final kernel binary.

For future fonts we will rely on the (not yet existing) file sytem.

However including 1 font directly is also useful for pancis, where we
can't rely on the fs working