Update rust version and X86 crate
The x86 crate forces us to use u64 instead of usize in many places.
This is probably a good thing, although I'm not 100% sure I like that we
now use u64 instead of usize internally in the mem module for sizes.
We'll see how that works in the long run. Everywhere else should still
use usize.
The rust update changes how `static mut` referesnces are accessed. This
is now done using the `addr_of` of `addr_of_mut` macro.