NVME: Work around for qemu nvme race condition
It looks like there is a bug in the qemu nvme sytem. They update the
phase before fully filling out the completion entry which means that we
read incomplete data. Therefor we drop the completion entry after
checking the phase and read it again. This seems to ensure that we get
the updated completion entry and not just a partial version with just
the phase set.