Mem: Fix bug allowing allocation of pages in page table
We did not remove all pages that are potentially part of the page table
from the set of possible page allocations in the kernel page allocator.

This could lead to 2 potential bugs when `page_alloc.alloc()` returns a
page that is within the page table:
 1. The page was not yet used by the page table. Later the page table
    starts to use the page table and overwrites the entries for this
    page. This leads to any number of pointers into the page table that
    should point somewhere else.
 2. The page table already uses the page. Trying to map the page will
    fail with a `PageInUse` Error.

We now make sure to exclude all potential page table pages from the
possible page allocations by the kernel.
Also added a test that check that those pages cannot be allocated.