FS: Move away from the INode terminology
It's a name used by linux, but I am not 100% sure this really fits
with my current design. Also no-one (even in the linux world) seems to
know what INode even stands for, therefor it is hard to translate this
to a new design without just being confusing.
I am sure the people in linux know what they mean and I am ok with using
the inode name within the fuse driver, but I don't want it in my kernel