FS: Start MemTree implementation
MemTree is a in memory view of the FileNode tree that is stored on disc.
So far I have implemented insert and some validity checks.

Conflicts:
Cargo.lock:
For some reason I used to require bootloader version 0.12.
However no such version exists (neither upstream or in my fork). I did
temporarially inc the version in a feature branch, however after a
discussion with upstream we decided that it was no real breaking change
so for now we are on the 0.11.x branch.
I still rely on my fork as there has not yet been a new release with my
feature branch included (it's just merged into main).