FS: FileId is now NonZero and not NonMax
Turns out that linux uses Inode 1 for the root of the fuse fs, therefor
I guess I will also use 1 for the root of my fs, not that it really
matters.
This also means I can now use 0 for `None` which also simplifies
debugging as `NonMax` is using bit manipulation which are hard to follow
in gdb.