FS: Fix root TreeNode Pointer not written to device
Turns out that the root pointer for the file node B-Tree was never
updated. Instead the initial value was just assumed to be correct.

When the root node is split this changed the location of the root node
on the device and therefor caused the fs to read an outdated tree, that
was incompatible with the directory tree, thereby causing not just 
data loss but also data corruption.