BlockDevice: Fix Block alignement issue
I always assumed that the `repr(align(size))` attribute would just
affect the alignment and not the size of the struct. However it turns
out this also effects the size, meaning that my old implementation of
`Block` was technically wrong and `Aligned` and `Block` are basically
identical.

This removes `Aligned` and fixes `Block`.
However this now means that it is no longer to create a `Block<T>` where
`size_of::<T>() > BLOCK_SIZE`.
This is something I kinda want anyway but did not know how to achieve
before.