CoreStatics: add way to force decrement interrupt count
this is usefull when interrupts exit early, e.g. in a taks system
a timer interrupt might be ended by an manuall iretq call instead
of the just exiting the x86-interrupt function normally.
In that case the AutoRefCounterGuard is never dropped so it is necessary
to manually decrement the interrupt counter